:green_heart:

+++ package/src/renderer/__tests__/e2e/PlatformBuffer.spec.tsx
@@ -1,10 +1,13 @@
 import { checkImage } from "../../../__tests__/setup";
-import { AlphaType, ColorType } from "../../../skia/types";
 import { setupSkia } from "../../../skia/__tests__/setup";
 import { surface } from "../setup";
 
 describe("Platform Buffers", () => {
   it("creates a platform buffer from an image", async () => {
+    // Skip outside iOS and Android
+    if (surface.OS === "web") {
+      return;
+    }
     const result = await surface.eval((Skia) => {
       const sur = Skia.Surface.Make(256, 256)!;
       const canvas = sur.getCanvas();
@@ -27,26 +30,27 @@ describe("Platform Buffers", () => {
   });
   it("creates an image from a platform buffer", async () => {
     const { Skia: Sk } = setupSkia();
+    // Skip outside iOS and Android
+    if (surface.OS === "web") {
+      return;
+    }
     // Skip test on Fabric (it runs on API Level 21 which doesn't support platform buffers)
     if (surface.arch === "fabric" && surface.OS === "android") {
       return;
     }
-    const result = await surface.eval(
-      (Skia) => {
-        const sur = Skia.Surface.Make(256, 256)!;
-        const canvas = sur.getCanvas();
-        canvas.drawColor(Skia.Color("cyan"));
-        const platformBuffer = Skia.Image.MakePlatformBuffer(
-          sur.makeImageSnapshot()
-        );
-        const image = Skia.Image.MakeImageFromPlatformBuffer(
-          platformBuffer.pointer
-        ).encodeToBytes();
-        platformBuffer.delete();
-        return Array.from(image);
-      },
-      { alphaType: AlphaType.Unpremul, colorType: ColorType.BGRA_8888 }
-    );
+    const result = await surface.eval((Skia) => {
+      const sur = Skia.Surface.Make(256, 256)!;
+      const canvas = sur.getCanvas();
+      canvas.drawColor(Skia.Color("cyan"));
+      const platformBuffer = Skia.Image.MakePlatformBuffer(
+        sur.makeImageSnapshot()
+      );
+      const image = Skia.Image.MakeImageFromPlatformBuffer(
+        platformBuffer.pointer
+      ).encodeToBytes();
+      platformBuffer.delete();
+      return Array.from(image);
+    });
     const image = Sk.Image.MakeImageFromEncoded(
       Sk.Data.fromBytes(new Uint8Array(result))
     )!;
